# backend/internal/httpd/apispec/openapi.yaml
openapi: 3.1.0
info:
  description: Loopback-only HTTP surface served by the Go daemon. Generated from
    Go (code-first) — do not edit by hand; run `go generate ./...`.
  title: Agent Orchestrator HTTP daemon
  version: 0.1.0-route-shell
servers:
- description: Local daemon (loopback only)
  url: http://127.0.0.1:3001
paths:
...
  /api/v1/questions/pending:
    get:
      operationId: listPendingDecisions
      parameters:
      - description: Filter to one or more of human_required, resolving, waiting_for_capacity.
          Omit for the default (human_required + resolving).
        in: query
        name: state
        schema:
          description: Filter to one or more of human_required, resolving, waiting_for_capacity.
            Omit for the default (human_required + resolving).
          enum:
          - human_required
          - resolving
          - waiting_for_capacity
          items:
            type: string
          type: array
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListWorkflowQuestionsResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Bad Request
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: List open/in-flight questions across all workflow runs (Checkpoint
        8K-B pass 3's global Pending Decisions inbox)
      tags:
      - decisions
...
  /api/v1/workflows/{workflowId}/questions:
    get:
      operationId: listWorkflowQuestions
      parameters:
      - description: Workflow run identifier.
        in: path
        name: workflowId
        required: true
        schema:
          description: Workflow run identifier.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListWorkflowQuestionsResponse'
          description: OK
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: List durable questions (Checkpoint 8K-A) captured for a workflow run
      tags:
      - workflows
  /api/v1/workflows/{workflowId}/questions/{questionId}:
    get:
      operationId: getWorkflowQuestion
      parameters:
      - description: Workflow run identifier.
        in: path
        name: workflowId
        required: true
        schema:
          description: Workflow run identifier.
          type: string
      - description: Workflow question identifier.
        in: path
        name: questionId
        required: true
        schema:
          description: Workflow question identifier.
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowQuestionResponseBody'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Get one durable question captured for a workflow run
      tags:
      - workflows
  /api/v1/workflows/{workflowId}/questions/{questionId}/answer:
    post:
      operationId: answerWorkflowQuestion
      parameters:
      - description: Workflow run identifier.
        in: path
        name: workflowId
        required: true
        schema:
          description: Workflow run identifier.
          type: string
      - description: Workflow question identifier.
        in: path
        name: questionId
        required: true
        schema:
          description: Workflow question identifier.
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnswerWorkflowQuestionRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowQuestionResponseBody'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Unprocessable Entity
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
          description: Not Implemented
      summary: Submit a human answer for a question that is awaiting one (Checkpoint
        8K-A)
      tags:
      - workflows
...
components:
  schemas:
    APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
      required:
      - error
      - code
      - message
      type: object
...
    AnswerWorkflowQuestionRequest:
      properties:
        choiceId:
          description: One of the question's structured choice ids.
          type:
          - "null"
          - string
        customText:
          description: Free-text human answer, when no structured choice applies.
          type:
          - "null"
          - string
      type: object
...
    ListWorkflowQuestionsResponse:
      properties:
        questions:
          items:
            $ref: '#/components/schemas/WorkflowQuestionResponse'
          type: array
      required:
      - questions
      type: object
...
    WorkflowQuestionChoiceResponse:
      properties:
        id:
          type: string
        label:
          type: string
      required:
      - id
      - label
      type: object
    WorkflowQuestionResponse:
      properties:
        answerSource:
          enum:
          - ""
          - policy
          - human
          - resolver
          type: string
        answerText:
          type: string
        answeredAt:
          type:
          - "null"
          - string
        askingHarness:
          type: string
        askingRole:
          type: string
        certainty:
          enum:
          - actual
          - inferred
          - unknown
          type: string
        classification:
          enum:
          - policy_resolvable
          - auto_resolvable
          - human_required
          - ambiguous
          type: string
        classificationReason:
          type: string
        createdAt:
          type: string
        delivered:
          type: boolean
        deliveredAt:
          type:
          - "null"
          - string
        id:
          type: string
        questionText:
          type: string
        resolverAdvisoryAnswer:
          type: string
        resolverEvidenceReferences:
          items:
            type: string
          type: array
        resolverHarness:
          type: string
        resolverProvider:
          type: string
        resolverReasonSummary:
          type: string
        sessionId:
          type: string
        state:
          enum:
          - pending
          - resolving
          - answered
          - human_required
          - cancelled
          type: string
        structuredChoices:
          items:
            $ref: '#/components/schemas/WorkflowQuestionChoiceResponse'
          type: array
        workflowRunId:
          type: string
        workflowStepId:
          type: string
      required:
      - id
      - workflowRunId
      - questionText
      - certainty
      - classification
      - state
      - createdAt
      - delivered
      type: object
    WorkflowQuestionResponseBody:
      properties:
        question:
          $ref: '#/components/schemas/WorkflowQuestionResponse'
      required:
      - question
      type: object
...